pipreqs_collection/requirements.txt
beautifulsoup4==4.12.2
dacite==1.6.0
geopandas==0.12.2
matplotlib==3.7.0
metar==1.9.0
ml_commons==2.3.0
networkx==3.0
numpy==1.24.2
pandas==1.5.3
ptflops==0.7
python_dateutil==2.8.2
PyYAML==6.0
rasterio==1.3.6
requests==2.28.2
scikit_learn==1.3.0
seaborn==0.11.2
sentinelsat==1.1.1
shapely==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.65.0
